requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-cors==0.7.0
aiosignal==1.3.2
astroid==3.3.8
attrs==25.1.0
bidict==0.23.1
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chart-studio==1.1.0
click==8.1.8
comtypes==1.1.14
contourpy==1.3.1
coverage==7.3.2
cryptography==44.0.1
cycler==0.12.1
dill==0.3.9
dnspython==2.7.0
# Editable install with no version control (eco-vehicle-modeling==0.1.0)
-e /Volumes/Learn_Space/eco_vehicle_project
elementpath==4.7.0
execnet==2.1.1
ezdxf==1.3.5
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.56.0
freezegun==1.2.2
frozenlist==1.5.0
graphviz==0.20.3
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
hypothesis==6.87.1
ibm-cloud-sdk-core==3.22.1
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm-watson==9.0.0
ibm_db==3.2.5
ibm_watson_machine_learning==1.0.367
ibmcloudant==0.10.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lomond==0.3.3
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mpmath==1.3.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
narwhals==1.28.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.1.4
pathspec==0.12.1
pillow==11.1.0
plantuml==0.3.0
plantuml-markdown==3.11.1
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
propcache==0.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyautocad==0.2.0
pycparser==2.22
pydot==3.0.4
pygraphviz==1.14
PyJWT==2.10.1
pylint==3.3.4
pymongo==4.11.1
pyparsing==3.2.1
PyQt6==6.4.2
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyqtgraph==0.13.7
pytest==8.3.4
pytest-aiohttp==1.1.0
pytest-asyncio==0.21.1
pytest-benchmark==5.1.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.3.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.2
retrying==1.3.4
rpds-py==0.23.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
simple-websocket==1.1.0
six==1.17.0
sortedcontainers==2.4.0
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.5.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xmlschema==3.4.3
yamllint==1.35.1
yarl==1.18.3
zipp==3.21.0
